Python/requirements.txt
networkx==3.0
coloredlogs==15.0.1
onnx==1.12.0
torch==1.13.0+cpu
numpy==1.24.1
graphviz==0.20.1
onnxruntime==1.13.1
tqdm==4.64.1
scikit-opt==0.6.6
PyYAML==6.0